third_party/blink/web_tests/fast/frames/sandboxed-iframe-attribute-parsing-06-expected.txt

CONSOLE ERROR: Error while parsing the 'sandbox' attribute: 'allowscripts' is an invalid sandbox flag.
CONSOLE ERROR: Blocked script execution in 'sandboxed-iframe-attribute-parsing-disallowed.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
Misspelling.

PASS
